Sending a form, receiving data from the server added.
