+++ js/upload-form/send-form.js
@@ -1,95 +1,31 @@
-import { setUserFormSubmit } from './validate-form.js';
-import { closeUserModal } from './show-form.js';
-import { isEscapeKey } from '../util.js';
+import { showFormSuccess, showFormError } from './show-form-message.js';
+import { pristine } from './validate-form.js';
+import { sendData } from '../server-api.js';
 
-const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
-const successMessage = successMessageTemplate.cloneNode(true);
-const successCloseButton = successMessage.querySelector('.success__button');
-
-const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
-const errorMessage = errorMessageTemplate.cloneNode(true);
-const errorCloseButton = errorMessage.querySelector('.error__button');
-
-const imageOverlay = document.querySelector('.img-upload__overlay');
+const form = document.querySelector('#upload-select-image');
+const submitButton = form.querySelector('.img-upload__submit');
 
 
-const closeFormSuccess = function () {
-  document.body.removeChild(successMessage);
-  document.removeEventListener('keydown', onFormSuccessEscKeydown);
-  document.removeEventListener('click', onFormSuccessWindowClick);
-  successCloseButton.removeEventListener('click', closeFormSuccess);
-};
-
-
-function onFormSuccessEscKeydown(evt) {
-  if (isEscapeKey(evt)) {
+const setUserFormSubmit = function (onSuccess, onFail) {
+  form.addEventListener('submit', (evt) => {
     evt.preventDefault();
-    closeFormSuccess();
-  }
-}
-
-
-function onFormSuccessWindowClick(evt) {
-  if (!evt.target.matches('.success__inner')) {
-    closeFormSuccess();
-  }
-}
-
-
-const closeFormError = function () {
-  document.body.removeChild(errorMessage);
-  document.removeEventListener('keydown', onFormErrorEscKeydown);
-  document.removeEventListener('click', onFormErrorWindowClick);
-  errorCloseButton.removeEventListener('click', closeFormError);
-  if (this === errorCloseButton) {
-    imageOverlay.classList.remove('hidden');
-  }
-  else {
-    closeUserModal();
-  }
+    const isValid = pristine.validate();
+    if (isValid) {
+      submitButton.disabled = true;
+      sendData(
+        () => {
+          onSuccess();
+          submitButton.disabled = false;
+        },
+        () => {
+          onFail();
+          submitButton.disabled = false;
+        },
+        new FormData(evt.target),
+      );
+    }
+  });
 };
 
-
-function onFormErrorEscKeydown(evt) {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closeFormError();
-  }
-}
-
-
-function onFormErrorWindowClick(evt) {
-  if (!evt.target.matches('.error__inner')) {
-    closeFormError();
-  }
-}
-
-
-const showFormMessage = function (element) {
-  document.body.appendChild(element);
-
-  if (element === successMessage) {
-    successCloseButton.addEventListener('click', closeFormSuccess);
-    document.addEventListener('keydown', onFormSuccessEscKeydown);
-    document.addEventListener('click', onFormSuccessWindowClick);
-  }
-  else if (element === errorMessage) {
-    errorCloseButton.addEventListener('click', closeFormError);
-    document.addEventListener('keydown', onFormErrorEscKeydown);
-    document.addEventListener('click', onFormErrorWindowClick);
-  }
-};
-
-
-const showFormSuccess = function () {
-  closeUserModal();
-  showFormMessage(successMessage);
-};
-
-
-const showFormError = function () {
-  imageOverlay.classList.add('hidden');
-  showFormMessage(errorMessage);
-};
 
 setUserFormSubmit(showFormSuccess, showFormError);
