+++ js/upload-form/validate-form.js
@@ -1,13 +1,11 @@
 import { countValueInArray, checkStringLength } from '../util.js';
-import { MAX_DESCRIPTION_LENGTH } from '../data.js';
-import { sendData } from '../server-api.js';
 
 const HASHTAG_RE = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
+const MAX_DESCRIPTION_LENGTH = 140;
 
 const form = document.querySelector('#upload-select-image');
 const formHashtags = form.querySelector('.text__hashtags');
 const formComment = form.querySelector('.text__description');
-const submitButton = form.querySelector('.img-upload__submit');
 
 
 const validateHashtags = function (value) {
@@ -25,7 +23,7 @@ const validateHashtags = function (value) {
 };
 
 
-const validateComment = function (comment) {
+const validateDescription = function (comment) {
   return checkStringLength(comment, MAX_DESCRIPTION_LENGTH);
 };
 
@@ -41,27 +39,6 @@ const pristine = new Pristine(form, {
 
 
 pristine.addValidator(formHashtags, validateHashtags, 'Неверный формат хэштегов.');
-pristine.addValidator(formComment, validateComment, 'Длина комментария не может составлять больше 140 символов.');
+pristine.addValidator(formComment, validateDescription, 'Длина комментария не может составлять больше 140 символов.');
 
-const setUserFormSubmit = function (onSuccess, onFail) {
-  form.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    const isValid = pristine.validate();
-    if (isValid) {
-      submitButton.disabled = true;
-      sendData(
-        () => {
-          onSuccess();
-          submitButton.disabled = false;
-        },
-        () => {
-          onFail('Не удалось отправить форму. Попробуйте ещё раз.', 5000);
-          submitButton.disabled = false;
-        },
-        new FormData(evt.target),
-      );
-    }
-  });
-};
-
-export { setUserFormSubmit, pristine };
+export { pristine };
